+++ lib/t/t_case9_save2psse.m
@@ -1,0 +1,67 @@
+function mpc = t_case9_save2psse
+%T_CASE9_SAVE2PSSE   Power flow data to test SAVE2PSSE.
+%   Please see CASEFORMAT for details on the case file format.
+
+%   MATPOWER
+
+%% MATPOWER Case Format : Version 2
+mpc.version = '2';
+
+%%-----  Power Flow Data  -----%%
+%% system MVA base
+mpc.baseMVA = 100;
+
+%% bus data
+%	bus_i	type	Pd	Qd	Gs	Bs	area	Vm	Va	baseKV	zone	Vmax	Vmin
+mpc.bus = [
+	1	3	0	0	0	0	1	1	0	345	1	1.1	0.9;
+	2	2	0	0	0	0	1	1	0	345	1	1.1	0.9;
+	3	2	0	0	0	0	1	1	0	345	1	1.1	0.9;
+	4	1	0	0	0	0	1	1	0	345	1	1.1	0.9;
+	5	1	90	30	0	0	1	1	0	345	1	1.1	0.9;
+	6	1	0	0	0	0	1	1	0	345	1	1.1	0.9;
+	7	1	100	35	0	0	1	1	0	345	1	1.1	0.9;
+	8	1	0	0	0	0	1	1	0	345	1	1.1	0.9;
+	9	1	125	50	0	0	1	1	0	345	1	1.1	0.9;
+];
+
+%% generator data
+%	bus	Pg	Qg	Qmax	Qmin	Vg	mBase	status	Pmax	Pmin	Pc1	Pc2	Qc1min	Qc1max	Qc2min	Qc2max	ramp_agc	ramp_10	ramp_30	ramp_q	apf
+mpc.gen = [
+	1	0	0	300	-300	1	100	1	250	90	0	0	0	0	0	0	0	0	0	0	0;
+	2	163	0	300	-300	1	100	1	300	10	0	0	0	0	0	0	0	0	0	0	0;
+	3	85	0	300	-300	1	100	1	270	10	0	0	0	0	0	0	0	0	0	0	0;
+];
+
+%% branch data
+%	fbus	tbus	r	x	b	rateA	rateB	rateC	ratio	angle	status	angmin	angmax
+mpc.branch = [
+	1	4	0	0.0576	0	250	250	250	0	0	1	-360	360;
+	4	5	0.017	0.092	0.158	250	250	250	0	0	1	-360	360;
+	5	6	0.039	0.17	0.358	150	150	150	0	0	1	-360	360;
+	3	6	0	0.0586	0	300	300	300	0	0	1	-360	360;
+	6	7	0.0119	0.1008	0.209	40	150	150	0	0	1	-360	360;
+	7	8	0.0085	0.072	0.149	250	250	250	0	0	1	-360	360;
+	8	2	0	0.0625	0	250	250	250	0	0	1	-360	360;
+	8	9	0.032	0.161	0	250	250	250	0.935	0	1	-360	360;
+	9	4	0.01	0.085	0	250	250	250	0	5	1	-360	360;
+];
+
+%%-----  DC Line Data  -----
+%	fbus	tbus	status	Pf	Pt	Qf	Qt	Vf	Vt	Pmin	Pmax	QminF	QmaxF	QminT	QmaxT	loss0	loss1
+mpc.dcline = [
+	3	4	1	4	4	0	0	1	1	3.4	4.6	1.0718	4.0448	1.0718	4.0448	0	0;
+	4	7	1	3	3	0	0	1	1	2.55	3.45	0.8038	3.0336	0.8038	3.0336	0	0;
+];
+
+mpc.bus_name = {
+            'bus 1       ';
+            'bus 2       ';
+            'bus 3       ';
+            'bus 4       ';
+            'bus 5       ';
+            'bus 6       ';
+            'bus 7       ';
+            'bus 8       ';
+            'bus 9       ';
+};