+++ CHANGES.md
@@ -15,6 +15,8 @@ Since last release
 #### 5/24/17
   - Remove OPF result columns from `gen` matrix returned by `psse2mpc`.
   - Limit length of name of star-buses added by `psse2mpc` to 12 chars.
+  - Add `save2psse` with support for exporting MATPOWER case data to
+    PSS/E RAW format.
 
 #### 5/23/17
   - Add support for `gentype` and `genfuel` fields of MATPOWER case struct